add plainRouter and remove legacy chat
