+++ web/src/features/support-chat/IntroSection.tsx
