+++ web/src/features/support-chat/SuccessSection.tsx
