+++ web/src/server/api/root.ts
@@ -40,9 +40,8 @@ import { dashboardWidgetRouter } from "./routers/dashboardWidgets";
 import { TableViewPresetsRouter } from "@/src/server/api/routers/tableViewPresets";
 import { automationsRouter } from "@/src/features/automations/server/router";
 import { defaultEvalModelRouter } from "@/src/features/evals/server/defaultEvalModelRouter";
-// import { plainRouter } from "@/src/features/support-chat/trpc/plain";
 import { slackRouter } from "@/src/features/slack/server/router";
-import { plainRouter } from "@/src/features/support-chat-2/trpc/router";
+import { plainRouter } from "@/src/features/support-chat/trpc/plainRouter";
 import { queueAssignmentRouter } from "@/src/features/annotation-queues/server/annotationQueueAssignments";
 import { surveysRouter } from "@/src/server/api/routers/surveys";
